fix(bench-v2): disambiguate ESP32 Wire address overload

+++ bench_v2/include/board.h
@@ -7,7 +7,7 @@ constexpr int SDA=27, SCL=22; // Legacy STEP/DIR unused in SPI motion mode.
 constexpr int SSI_CLK[2]={14,13}, SSI_DATA[2]={34,35};
 constexpr int ESTOP=39; // External 10k pull-up; NC contact to GND. No internal pull-up!
 constexpr int SOL_A=25, SOL_B=26, SOL_EN=15, SOL_HALL[2]={32,33};
-constexpr unsigned MCP_ADDR=0x20;
+constexpr uint8_t MCP_ADDR=0x20;
 inline void safeOutputs() {
     for(int i=0;i<2;++i) { digitalWrite(EN[i],HIGH); pinMode(EN[i],OUTPUT); digitalWrite(CS[i],HIGH); pinMode(CS[i],OUTPUT); }
     digitalWrite(SOL_EN,LOW); pinMode(SOL_EN,OUTPUT);
